[Decomp][BattleGround] Cohesion split: extract rewards + object spawning to BattleGround*.cpp
